+++ engine/crates/engine/src/system/api/mod.rs
@@ -13,6 +13,7 @@ mod object_api;
 mod render_command_api;
 mod scene_api;
 mod time_api;
+mod asset_load_api;
 
 pub use asset_api::AssetApi;
 pub use entity_api::EntityApi;
@@ -21,3 +22,4 @@ pub use object_api::ObjectApi;
 pub use render_command_api::RenderCommandApi;
 pub use scene_api::SceneCommandApi;
 pub use time_api::TimeApi;
+pub use asset_load_api::AssetLoadApi;